# swagger.json
{
	"swagger": "2.0",
	"info": {
		"description": "Api Documentation",
		"version": "1.0",
		"title": "Api Documentation",
		"termsOfService": "urn:tos",
		"contact": {
			
		},
		"license": {
			"name": "Apache 2.0",
			"url": "http://www.apache.org/licenses/LICENSE-2.0"
		}
	},
	"host": "localhost:8080",
	"basePath": "/",
...
	"paths": {
		"/books": {
			"get": {
				"tags": [
					"rest-api-controller"
				],
				"summary": "getAllBooks",
				"operationId": "getAllBooksUsingGET",
				"produces": [
					"application/xml",
					"application/json"
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/Books"
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				}
			},
			"post": {
				"tags": [
					"rest-api-controller"
				],
				"summary": "addBooks",
				"operationId": "addBooksUsingPOST",
				"consumes": [
					"application/xml",
					"application/json"
				],
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"in": "body",
						"name": "bookList",
						"description": "bookList",
						"required": true,
						"schema": {
							"$ref": "#/definitions/Books"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "string"
						}
					},
					"201": {
						"description": "Created"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				}
			}
		},
...
	},
	"definitions": {
		"Book": {
			"type": "object",
			"properties": {
				"bName": {
					"type": "string"
				},
				"bookPrice": {
					"type": "number",
					"format": "double"
				},
				"id": {
					"type": "integer",
					"format": "int32"
				}
			},
			"title": "Book"
		},
		"Books": {
			"type": "object",
			"properties": {
				"bookList": {
					"type": "array",
					"items": {
						"$ref": "#/definitions/Book"
					}
				}
			},
			"title": "Books"
		},
...
	}
}